PEAR-424 - move cohort message code up component tree so it triggers on projects page
