+++ packages/portal-proto/src/features/cohortBuilder/ContextBar.tsx
@@ -1,18 +1,9 @@
 import React, { useEffect, useState } from "react";
-import { showNotification } from "@mantine/notifications";
 import { CollapsibleContainer } from "@/components/CollapsibleContainer";
 import { Tabs } from "@mantine/core";
 import { ContextualCasesView } from "../cases/CasesView/CasesView";
 import CountButton from "./CountButton";
 import CohortManager from "./CohortManager";
-import {
-  DeleteCohortNotification,
-  DiscardChangesCohortNotification,
-  ErrorCohortNotification,
-  NewCohortNotification,
-  SavedCohortNotification,
-  NewCohortNotificationWithSetAsCurrent,
-} from "@/features/cohortBuilder/CohortNotifications";
 import {
   useCoreDispatch,
   useCoreSelector,
@@ -20,9 +11,6 @@ import {
   DEFAULT_COHORT_ID,
   selectCurrentCohortId,
   setActiveCohort,
-  selectCohortMessage,
-  selectCurrentCohortName,
-  clearCohortMessage,
   setCohortList,
   useGetCohortsByContextIdQuery,
   buildGqlOperationToFilterSet,
@@ -100,93 +88,6 @@ const ContextBar: React.FC = () => {
   const currentCohortId = useCoreSelector((state) =>
     selectCurrentCohortId(state),
   );
-  const currentCohortName = useCoreSelector((state) =>
-    selectCurrentCohortName(state),
-  );
-  const cohortMessage = useCoreSelector((state) => selectCohortMessage(state));
-
-  useEffect(() => {
-    if (cohortMessage) {
-      const cmdAndParam = cohortMessage.split("|", 3);
-      if (cmdAndParam.length == 3) {
-        if (cmdAndParam[0] === "newCohort") {
-          showNotification({
-            message: <NewCohortNotification cohortName={cmdAndParam[1]} />,
-            classNames: {
-              description: "flex flex-col content-center text-center",
-            },
-            autoClose: 5000,
-          });
-        }
-        if (cmdAndParam[0] === "deleteCohort") {
-          showNotification({
-            message: <DeleteCohortNotification cohortName={cmdAndParam[1]} />,
-            classNames: {
-              description: "flex flex-col content-center text-center",
-            },
-            autoClose: 5000,
-          });
-        }
-        if (cmdAndParam[0] === "savedCohort") {
-          showNotification({
-            message: <SavedCohortNotification />,
-            classNames: {
-              description: "flex flex-col content-center text-center",
-            },
-            autoClose: 5000,
-          });
-        }
-        if (cmdAndParam[0] === "discardChanges") {
-          showNotification({
-            message: <DiscardChangesCohortNotification />,
-            classNames: {
-              description: "flex flex-col content-center text-center",
-            },
-            autoClose: 5000,
-          });
-        }
-        if (cmdAndParam[0] === "error") {
-          showNotification({
-            message: <ErrorCohortNotification errorType={cmdAndParam[1]} />,
-            classNames: {
-              description: "flex flex-col content-center text-center",
-            },
-            autoClose: 5000,
-          });
-        }
-        if (cmdAndParam[0] === "newCasesCohort") {
-          showNotification({
-            message: (
-              <NewCohortNotificationWithSetAsCurrent
-                cohortName={cmdAndParam[1]}
-                cohortId={cmdAndParam[2]}
-              />
-            ),
-            classNames: {
-              description: "flex flex-col content-center text-center",
-            },
-            autoClose: 5000,
-          });
-        }
-        if (cmdAndParam[0] === "newProjectsCohort") {
-          showNotification({
-            message: (
-              <NewCohortNotificationWithSetAsCurrent
-                cohortName={cmdAndParam[1]}
-                cohortId={cmdAndParam[2]}
-              />
-            ),
-            classNames: {
-              description: "flex flex-col content-center text-center",
-            },
-            autoClose: 5000,
-          });
-        }
-      }
-
-      coreDispatch(clearCohortMessage());
-    }
-  }, [cohortMessage, coreDispatch, currentCohortName]);
 
   useEffect(() => {
     setCurrentIndex(currentCohortId);
