+++ packages/portal-proto/src/features/layout/UserFlowVariedPages.tsx
@@ -1,4 +1,5 @@
 import { PropsWithChildren, ReactNode, useEffect } from "react";
+import { showNotification } from "@mantine/notifications";
 import {
   isString,
   useCoreSelector,
@@ -6,9 +7,20 @@ import {
   fetchNotifications,
   selectBanners,
   fetchUserDetails,
+  selectCurrentCohortName,
+  selectCohortMessage,
+  clearCohortMessage,
 } from "@gff/core";
 import Banner from "@/components/Banner";
 import { Button } from "@mantine/core";
+import {
+  DeleteCohortNotification,
+  DiscardChangesCohortNotification,
+  ErrorCohortNotification,
+  NewCohortNotification,
+  SavedCohortNotification,
+  NewCohortNotificationWithSetAsCurrent,
+} from "@/features/cohortBuilder/CohortNotifications";
 import { Header } from "./Header";
 import { Footer } from "./Footer";
 
@@ -35,6 +47,94 @@ export const UserFlowVariedPages: React.FC<UserFlowVariedPagesProps> = ({
   }, []);
 
   const banners = useCoreSelector((state) => selectBanners(state));
+
+  const currentCohortName = useCoreSelector((state) =>
+    selectCurrentCohortName(state),
+  );
+  const cohortMessage = useCoreSelector((state) => selectCohortMessage(state));
+
+  useEffect(() => {
+    if (cohortMessage) {
+      const cmdAndParam = cohortMessage.split("|", 3);
+      if (cmdAndParam.length == 3) {
+        if (cmdAndParam[0] === "newCohort") {
+          showNotification({
+            message: <NewCohortNotification cohortName={cmdAndParam[1]} />,
+            classNames: {
+              description: "flex flex-col content-center text-center",
+            },
+            autoClose: 5000,
+          });
+        }
+        if (cmdAndParam[0] === "deleteCohort") {
+          showNotification({
+            message: <DeleteCohortNotification cohortName={cmdAndParam[1]} />,
+            classNames: {
+              description: "flex flex-col content-center text-center",
+            },
+            autoClose: 5000,
+          });
+        }
+        if (cmdAndParam[0] === "savedCohort") {
+          showNotification({
+            message: <SavedCohortNotification />,
+            classNames: {
+              description: "flex flex-col content-center text-center",
+            },
+            autoClose: 5000,
+          });
+        }
+        if (cmdAndParam[0] === "discardChanges") {
+          showNotification({
+            message: <DiscardChangesCohortNotification />,
+            classNames: {
+              description: "flex flex-col content-center text-center",
+            },
+            autoClose: 5000,
+          });
+        }
+        if (cmdAndParam[0] === "error") {
+          showNotification({
+            message: <ErrorCohortNotification errorType={cmdAndParam[1]} />,
+            classNames: {
+              description: "flex flex-col content-center text-center",
+            },
+            autoClose: 5000,
+          });
+        }
+        if (cmdAndParam[0] === "newCasesCohort") {
+          showNotification({
+            message: (
+              <NewCohortNotificationWithSetAsCurrent
+                cohortName={cmdAndParam[1]}
+                cohortId={cmdAndParam[2]}
+              />
+            ),
+            classNames: {
+              description: "flex flex-col content-center text-center",
+            },
+            autoClose: 5000,
+          });
+        }
+        if (cmdAndParam[0] === "newProjectsCohort") {
+          showNotification({
+            message: (
+              <NewCohortNotificationWithSetAsCurrent
+                cohortName={cmdAndParam[1]}
+                cohortId={cmdAndParam[2]}
+              />
+            ),
+            classNames: {
+              description: "flex flex-col content-center text-center",
+            },
+            autoClose: 5000,
+          });
+        }
+      }
+
+      dispatch(clearCohortMessage());
+    }
+  }, [cohortMessage, dispatch, currentCohortName]);
 
   return (
     <div className="flex flex-col min-h-screen min-w-full bg-base-max">
